(feat) move liquid to new location
