+++ hummingbot/connector/exchange/liquid/liquid_order_book_message.py
