+++ hummingbot/strategy/pure_market_making/start.py
@@ -10,7 +10,7 @@ from hummingbot.strategy.pure_market_making import (
     APIAssetPriceDelegate
 )
 from hummingbot.strategy.pure_market_making.pure_market_making_config_map import pure_market_making_config_map as c_map
-from hummingbot.market.paper_trade import create_paper_trade_market
+from hummingbot.connector.exchange.paper_trade import create_paper_trade_market
 from hummingbot.connector.exchange_base import ExchangeBase
 from decimal import Decimal
