+++ hummingbot/user/user_balances.py
@@ -3,7 +3,7 @@ from hummingbot.market.bittrex.bittrex_market import BittrexMarket
 from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_market import CoinbaseProMarket
 from hummingbot.market.huobi.huobi_market import HuobiMarket
 from hummingbot.market.kucoin.kucoin_market import KucoinMarket
-from hummingbot.market.liquid.liquid_market import LiquidMarket
+from hummingbot.connector.exchange.liquid.liquid_market import LiquidMarket
 from hummingbot.market.kraken.kraken_market import KrakenMarket
 from hummingbot.market.eterbase.eterbase_market import EterbaseMarket
 from hummingbot.connector.exchange.crypto_com.crypto_com_exchange import CryptoComExchange
